bitunix-trading-service/requirements.txt
# Bitunix Cryptocurrency Trading Service Dependencies

# Core Dependencies
asyncpg>=0.28.0             # PostgreSQL async driver
python-dotenv>=1.0.0        # Environment variable management
requests>=2.31.0            # HTTP requests for Bitunix API
aiohttp>=3.9.0              # Async HTTP client (alternative)

# Data Processing
pandas>=2.0.0               # Data analysis (optional)

# Cryptocurrency Trading
ccxt>=4.0.0                 # Unified crypto exchange library (optional)
websocket-client>=1.6.0     # WebSocket for real-time data (optional)

# Security
cryptography>=41.0.0        # For API signatures and secure credentials

# Utilities
colorlog>=6.7.0             # Colored logging output
python-dateutil>=2.8.2     # Date parsing utilities

# Development Dependencies (optional)
pytest>=7.4.0               # Testing framework
pytest-asyncio>=0.21.0     # Async testing support
black>=23.0.0               # Code formatting
flake8>=6.0.0               # Code linting

# Performance Monitoring (optional)
psutil>=5.9.0               # System monitoring 